# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden:
    <ul>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird ein JSON-Array von JSON-Objekten, wobei ein JSON-Objekt eine Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Bei einer gültigen <code>DELETE</code>-Anfrage wird der Code <code>204</code> zurückgegeben.</b> Ist die Ressource nicht vorhanden, wird dies mit Code <code>404</code> belegt.</li>
...
      <li><b>Antworten auf gültige <code>POST</code>- und <code>PATCH</code>-Anfragen beinhalten nichts.</b></li>
...
      <li><b>Die erwähnten IDs sind RowIDs!!!</b></li>
    </ul>
...
servers:
...
paths:
...
  /ausleihen:
    post:
      tags:
        - Authentifizierter und als Kunde authorisierter Anwender
      security:
        - basicAuth: []
      summary: Fügt eine (eigene) Ausleihe hinzu.
      description: >-
        Eine **Ausleihe** wird mit den entsprechenden Werten hinzugefügt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    exemplarid:
                      type: integer
                      description: '**RowID** des Exemplars.'
                  required:
                    - exemplarid
                - $ref: '#/components/schemas/ausleihe'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
    get:
      tags:
        - Authentifizierter und als Kunde authorisierter Anwender
      security:
        - basicAuth: []
      summary: Gibt (eigene) Ausleihen zurück.
      description: >-
        Eigene **Ausleihen** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: zurueckgegeben
          in: query
          description: 'Nur Ausleihen, deren **Rückgabe-Status** `zurueckgegeben` ist, werden zurückgegeben.'
          schema:
            type: boolean
        - name: beginn
          in: query
          description: 'Nur Ausleihen, deren **Beginn** <=  `beginn` ist, werden zurückgegeben.'
          schema:
            type: string
            format: date
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        ausleiheid:
                          type: integer
                          description: '**RowID der Ausleihe**.'
                        exemplarid:
                          type: integer
                          description: '**RowID des Exemplars**.'
                      required:
                        - exemplarid
                        - ausleiheid
                    - $ref: '#/components/schemas/ausleihe'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
  /adressen:
    get:
      tags:
        - Anwender
      summary: Gibt Adressen zurück.
      description: >-
        Alle **Adressen** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: hausnummer
          in: query
          description: 'Nur Adressen, deren **Hausnummer** `hausnummer` enthält, werden zurückgegeben.'
          schema:
            type: string
        - name: strasse
          in: query
          description: 'Nur Adressen, deren **Strasse** `strasse` enthält, werden zurückgegeben.'
          schema:
            type: string
        - name: plz
          in: query
          description: 'Nur Adressen, deren **PLZ** `plz` enthält, werden zurückgegeben.'
          schema:
            type: string
        - name: stadt
          in: query
          description: 'Nur Adressen, deren **Stadt** `stadt` enthält, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        adresseid:
                          type: integer
                          description: '**RowID der Adresse**.'
                      required:
                        - adresseid
                    - $ref: '#/components/schemas/adresse'
    patch:
      tags:
        - Authentifizierter und als Kunde authorisierter Anwender
      security:
        - basicAuth: []
      summary: Ändert die (eigene) Adresse des Kunden.
      description: >-
        Die **Adresse** vom authentifizierten Kunden wird gemäß der Query-Parameter geändert.
      parameters:
        - name: hausnummer
          in: query
          description: 'Die neue **Hausnummer**.'
          schema:
            type: string
        - name: strasse
          in: query
          description: 'Die neue **Strasse**.'
          schema:
            type: string
        - name: stadt
          in: query
          description: 'Die neue **Stadt**.'
          schema:
            type: string
        - name: plz
          in: query
          description: 'Die neue **PLZ**.'
          schema:
            type: string
      responses:
        '204':
          $ref: '#/components/responses/204'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'            
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource.
          schema:
            type: string
    '204':
      description: '**No Content**.'
...
    '400':
      description: '**Bad Request**.'
    '401':
      description: '**Unauthorized**.'
      headers:
        WWW-Authenticate:
          schema:
            type: string
    '403':
      description: '**Forbidden**.'
...
  schemas:
...
    ausleihe:
      type: object
      properties:
        zurueckgegeben:
          type: boolean
          description: '**Rückgabe-Status** der Ausleihe. Nicht zurückgegeben = false / Zurückgegeben = true.'
        beginn:
          type: string
          format: date
          description: '**Ausleihbeginn** im Format "YYYY-MM-DD".'
        ende:
          type: string
          format: date
          description: '**Ausleihende** im Format "YYYY-MM-DD".'
      required:
        - zurueckgegeben
        - beginn
        - ende
...
    adresse:
      type: object
      properties:
        plz:
          type: string
          description: '**PLZ** der Adresse.'
        stadt:
          type: string
          description: '**Stadt** der Adresse.'
        strasse:
          type: string
          description: '**Straße** der Adresse.'
        hausnummer:
          type: string
          description: '**Hausnummer** der Adresse.'
      required:
        - plz
        - stadt
        - strasse
        - hausnummer